+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.2.2
 - Removed the width limit on slider input field
+- Improved Color Picker by adding RGB sliders and hex input
 ## 1.2.1
 - Fixed a bug where the phone wasn't closing with Tab, only Esc after using a slider
 - Fixed compatibility bug where LogManager would crash the Melon init on ML 0.7.2